user location nearest beacon

+++ app/src/main/java/com/example/myway/Model/BluetoothRep.java
@@ -25,7 +25,7 @@ public class BluetoothRep {
         if (rssi==0){
             return Integer.MAX_VALUE;
         }
-        float temp = (RSSI_AT_ONE_METER - rssi)/(10*ENVIROMENTAL_FACTOR);
+        double temp = (RSSI_AT_ONE_METER - rssi)/(10*ENVIROMENTAL_FACTOR);
         d=Math.pow(10,temp);
         return d;
     }
